fix(react-ui-components): change copy-icons script to generate svg symbols on the correct path
When the package manager defined on the project is `pnpm` the node_modules are symlinks. Since this script uses relative paths, this script didn't work as pretended. To overpass this problem the script now has `CUSTOM_INST` environment variable that will copy the icons based on the project directory calls. Check README notes.